+++ exports.ts
@@ -1,7 +1,10 @@
 /** common exports go in this module */
 
 export {BigNumber} from "bignumber.js";
-export {BaseClient} from './src/BaseClient';
+export {
+    BaseClient, ClientConfig, Signer, SigningOpts, PubKeyAndSigner, Operator,
+    PrivateKey, Nodes, Node
+}from './src/BaseClient';
 
 export {AccountCreateTransaction} from './src/account/AccountCreateTransaction';
 export {CryptoTransferTransaction} from './src/account/CryptoTransferTransaction';
@@ -12,3 +15,5 @@ export {
     Ed25519PrivateKey, Ed25519PublicKey, ThresholdKey, PublicKey,
     generateMnemonic, KeyMismatchException
 } from './src/Keys';
+
+export * from './src/typedefs';
