+++ src/TransactionBuilder.ts
@@ -1,4 +1,4 @@
-import {AccountId, BaseClient, Node, TransactionId} from "./BaseClient";
+import {BaseClient, Node} from "./BaseClient";
 import {TransactionBody} from "./generated/TransactionBody_pb";
 import {getProtoAccountId, getProtoTxnId, newDuration, newTxnId} from "./util";
 import {Transaction} from "./Transaction";
@@ -6,6 +6,8 @@ import {Transaction as Transaction_} from "./generated/Transaction_pb";
 import {grpc} from "@improbable-eng/grpc-web";
 import {TransactionResponse} from "./generated/TransactionResponse_pb";
 import BigNumber from "bignumber.js";
+
+import {AccountId, TransactionId} from "./typedefs";
 import UnaryMethodDefinition = grpc.UnaryMethodDefinition;
 
 /**
