+++ src/account/CryptoTransferTransaction.ts
@@ -2,7 +2,7 @@ import {TransactionBuilder} from "../TransactionBuilder";
 import {Transaction} from "../generated/Transaction_pb";
 import {TransactionResponse} from "../generated/TransactionResponse_pb";
 import {grpc} from "@improbable-eng/grpc-web";
-import {AccountId, BaseClient} from "../BaseClient";
+import {BaseClient} from "../BaseClient";
 import {
     AccountAmount,
     CryptoTransferTransactionBody,
@@ -11,6 +11,8 @@ import {
 import {checkNumber, getProtoAccountId, reqDefined} from "../util";
 import BigNumber from "bignumber.js";
 import {CryptoService} from "../generated/CryptoService_pb_service";
+
+import {AccountId} from "../typedefs";
 
 export class CryptoTransferTransaction extends TransactionBuilder {
     private readonly body: CryptoTransferTransactionBody;
